requirements.txt
cryptography==42.0.8
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.7.3
pydantic_core==2.18.4
PyMySQL==1.1.1
pytest==8.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
stripe==9.10.0
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
uvicorn==0.30.1
python-dotenv==1.0.1
python-jose==3.3.0
alembic==1.13.2